+++ dashboard/src/components/players/PlayerDetail.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
-import { playerService, parentService } from '../../lib/supabase';
+import { playerService, parentService } from '../../../../src/lib/supabase';
 
 const PlayerDetail = () => {
   const { id } = useParams();
